perf(stratego-trainer): bf16 training (mixed precision) behind --bf16-train
Training forward/backward matmuls run in bf16 (~1.75x), with fp32 master weights (grads cast bf16->fp32 before the optimizer) and the loss math forced to fp32 — cast the net's logits/value_logp/ent_pred to fp32 (the unbounded ent_pred head overflows in bf16; that was the immediate-NaN cause). Setup baseline kept fp32 (precise frozen reference). Smoke: train pass ~4.3s->0.86s, learning (WSrand 0.67 by iter 17), only the same early-warmup transient the fp32 path also self-heals. Off by default pending the full danger-zone validation.

+++ ml/stratego-trainer/stratego_trainer/move_loss.py
@@ -82,7 +82,7 @@ def move_loss_and_stats(net, batch, magnet_coef, cfg):
     out = net(obs, legal_mask=legal)
     move_logits = out["move_logits"].astype(mx.float32)
     log_probs = move_logits - mx.logsumexp(move_logits, axis=-1, keepdims=True)  # (B,1800)
-    value_logp = out["value_logp"]  # (B,3) log-softmax
+    value_logp = out["value_logp"].astype(mx.float32)  # (B,3) log-softmax
 
     # PPO-clip policy loss on the chosen action. Clamp the log-ratio before exp so
     # a policy that drifts far from the data policy can't overflow exp() to inf
